requirements.txt
Pillow==9.1.0
tqdm==4.64.0